+++ Pandora/Paradigm/Structure/Stream.hs
@@ -14,7 +14,7 @@ import Pandora.Paradigm.Primary.Functor.Identity (Identity (Identity))
 import Pandora.Paradigm.Primary.Functor.Wye (Wye (Left, Right))
 import Pandora.Paradigm.Primary.Transformer.Construction (Construction (Construct), deconstruct, (.-+))
 import Pandora.Paradigm.Primary.Transformer.Tap (Tap (Tap))
-import Pandora.Paradigm.Structure.Ability.Rotatable (Rotatable (Rotational, rotation), rotate)
+import Pandora.Paradigm.Structure.Ability.Convertible (Convertible (Conversion, conversion), convert)
 import Pandora.Paradigm.Structure.Ability.Zipper (Zipper)
 import Pandora.Paradigm.Schemes.TU (TU (TU), type (<:.>))
 
@@ -22,19 +22,19 @@ type Stream = Construction Identity
 
 type instance Zipper Stream = Tap (Delta <:.> Stream)
 
-instance Rotatable Left (Tap (Delta <:.> Stream)) where
-	type Rotational Left (Tap (Delta <:.> Stream)) = Tap (Delta <:.> Stream)
-	rotation (extract . run -> Tap x (TU (bs :^: fs))) = Tap (extract bs) . TU
+instance Convertible Left (Tap (Delta <:.> Stream)) where
+	type Conversion Left (Tap (Delta <:.> Stream)) = Tap (Delta <:.> Stream)
+	conversion (extract . run -> Tap x (TU (bs :^: fs))) = Tap (extract bs) . TU
 		$ extract (deconstruct bs) :^: Construct x (point fs)
 
-instance Rotatable Right (Tap (Delta <:.> Stream)) where
-	type Rotational Right (Tap (Delta <:.> Stream)) = Tap (Delta <:.> Stream)
-	rotation (extract . run -> Tap x (TU (bs :^: fs))) = Tap (extract fs) . TU
+instance Convertible Right (Tap (Delta <:.> Stream)) where
+	type Conversion Right (Tap (Delta <:.> Stream)) = Tap (Delta <:.> Stream)
+	conversion (extract . run -> Tap x (TU (bs :^: fs))) = Tap (extract fs) . TU
 		$ Construct x (point bs) :^: extract (deconstruct fs)
 
 instance {-# OVERLAPS #-} Extendable (Tap (Delta <:.> Stream)) where
 	z =>> f = let move rtt = extract . deconstruct $ point . rtt .-+ z
-		in f <$> Tap z (TU $ move (rotate @Left) :^: move (rotate @Right))
+		in f <$> Tap z (TU $ move (convert @Left) :^: move (convert @Right))
 
 repeat :: a :=> Stream
 repeat x = Construct x . Identity $ repeat x