+++ CHANGELOG.md
@@ -339,3 +339,4 @@
 * Define experimental `adjust`, `magnify`, `=<>`, `~<>`
 * Define experimental `Accessible` type class
 * Rename `access` method to `look` in `Optics` module
+* Remove `Rotatable` type class in favor of `Convertible`